refactor: 한 라운드 승자 화살표 그리는 함수(playOnce) 로직 수정(thanks yechoi)
